refactor: mover serviço para application/service
