+++ src/main/java/com/sensedia/sample/consents/rest/ConsentApi.java
@@ -4,7 +4,7 @@ import com.sensedia.sample.consents.dto.ConsentRequestDTO;
 import com.sensedia.sample.consents.dto.ConsentResponseDTO;
 import com.sensedia.sample.consents.dto.ConsentUpdateDTO;
 import com.sensedia.sample.consents.dto.PageResponseDTO;
-import com.sensedia.sample.consents.service.ConsentService;
+import com.sensedia.sample.consents.application.service.ConsentService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
