+++ src/test/java/com/sensedia/sample/consents/service/impl/ConsentServiceImplTest.java
@@ -1,5 +1,6 @@
 package com.sensedia.sample.consents.service.impl;
 
+import com.sensedia.sample.consents.application.service.impl.ConsentServiceImpl;
 import com.sensedia.sample.consents.client.GitHubClient;
 import com.sensedia.sample.consents.domain.model.Consent;
 import com.sensedia.sample.consents.domain.enums.ConsentStatus;
